Update Web Camera using Python OpenCV.py
OpenCV is fast and customizable. So one can use it for real time image processing. Thus OpenCV comes out to be very useful for using in image processing Robotics applications. In many image processing based robotics applications, a camera is mounted in robot. Using input from mounted camera behavior of robot is controlled.

In this code we will access the camera of laptop or any webcam using opencv and produce live video stream.